+++ src/filters/reader/CDDAReader/CDDAReader.cpp
@@ -474,7 +474,7 @@ HRESULT CCDDAStream::Read(PBYTE pbBuffer, DWORD dwBytesToRead, BOOL bAlign, LPDW
     }
 
     if (pdwBytesRead) {
-        *pdwBytesRead = pbBuffer - pbBufferOrg;
+        *pdwBytesRead = DWORD(pbBuffer - pbBufferOrg);
     }
     m_llPosition += pbBuffer - pbBufferOrg;
